refactored Main to include sub-component for Travellers
